Consolidate Rogue mod entrypoint
